chore: move generateUserTopWrapper
